src/Day 5/Image recognizing bot/requirements.txt
discord.py==1.7.3
keras==2.7.0
matplotlib==3.5.1
numpy==1.22.1
pandas==1.3.5
Pillow==9.0.0
requests==2.26.0
